chore: cast floating to float

+++ bindings/python/test/solver/test_numerical_solver.py
@@ -133,7 +133,7 @@ class TestNumericalSolver:
         assert 5e-9 >= abs(state_vector[0] - math.sin(integration_duration))
         assert 5e-9 >= abs(state_vector[1] - math.cos(integration_duration))
 
-        integration_durations = np.arange(100.0, 200.0, 50.0)
+        integration_durations = np.arange(100.0, 200.0, 50.0).astype(float)
         solutions = numerical_solver.integrate_duration(
             initial_state_vec, integration_durations, oscillator
         )
@@ -157,7 +157,7 @@ class TestNumericalSolver:
         assert 5e-9 >= abs(state_vector[0] - math.sin(end_time))
         assert 5e-9 >= abs(state_vector[1] - math.cos(end_time))
 
-        end_times = np.arange(600.0, 700.0, 50.0)
+        end_times = np.arange(600.0, 700.0, 50.0).astype(float)
         solutions = numerical_solver.integrate_time(
             initial_state_vec, start_time, end_times, oscillator
         )
